refactor: shadcn, radix 관련은 src/ui 디렉토리에
