+++ src/app/_shared/hooks/useAppNavigation.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 import { useRouter } from "next/navigation";
 
-import useNavigationStore from "@/app/stores/useNavigationStore";
+import useNavigationStore from "@/app/_shared/stores/useNavigationStore";
 
 import { NavigationPage } from "@/lib/types/navgationType";
