+++ src/app/_shared/stores/useNavigationStore.tsx
