+++ src/ui/form.tsx
@@ -12,7 +12,7 @@ import {
 import * as LabelPrimitive from "@radix-ui/react-label"
 import { Slot } from "@radix-ui/react-slot"
 
-import { Label } from "@/app/components/ui/label"
+import { Label } from "@/ui/label"
 
 import { cn } from "@/lib/utils"
